remove pointless/incorrect cast. ln_Pri is a BYTE in both nodes.
git-svn-id: https://svn.aros.org/svn/aros/trunk@53259 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/SysMon/tasks.c
@@ -38,7 +38,7 @@ VOID RefreshTask(struct TaskInfo *ti)
 {
     /* Cache values we need incase something happens to the task .. */
     ti->TINode.ln_Type = ti->Task->tc_Node.ln_Type;
-    ti->TINode.ln_Pri = (WORD)ti->Task->tc_Node.ln_Pri;
+    ti->TINode.ln_Pri = ti->Task->tc_Node.ln_Pri;
     ti->Flags |= TIF_ENABLED;
 }
